+++ CMakeLists.txt
@@ -426,9 +426,9 @@ set(CORE_SOURCE
         src/backend/PatchRecord.cpp
         src/backend/PatchRecordComponent.cpp
         src/backend/Writable.cpp
-        src/backend/ScientificDefaults.cpp
-        src/backend/ScientificDefaults_auxiliary.cpp
-        src/backend/ScientificDefaults_impl.cpp
+        src/backend/scientific_defaults/ScientificDefaults.cpp
+        src/backend/scientific_defaults/ScientificDefaults_auxiliary.cpp
+        src/backend/scientific_defaults/ScientificDefaults_impl.cpp
         src/auxiliary/OneDimensionalBlockSlicer.cpp
         src/helper/list_series.cpp
         src/snapshots/ContainerImpls.cpp
