+++ include/openPMD/backend/scientific_defaults/ScientificDefaults_auxiliary.hpp
