+++ src/backend/PatchRecord.cpp
@@ -20,7 +20,7 @@
  */
 #include "openPMD/backend/PatchRecord.hpp"
 #include "openPMD/auxiliary/Memory.hpp"
-#include "openPMD/backend/ScientificDefaults.hpp"
+#include "openPMD/backend/scientific_defaults/ScientificDefaults.hpp"
 
 #include <iostream>
